Added a sink module for rate-limiting
Created a separate sink module for rate-limiting, and incorporated it
into `rec.c`. In `rec.c`, it creates the rate-limit sink based on
whether or not rate-limiting is enabled. If so, it makes the `log_sink`
a rate-limit sink with a parameter of what would have been the log_sink.
Program writes to rate-limit sink to write to log_sink and log.

In `rate_limit_sink`, added all of the necessary methods present in the
other sinks (is_valid, cleanup, init, write, cut, flush). Calculation of
rate is using packet timestamps with no checks for extra things (i.e.
encoding, writing to the log, etc). Added sub-components to the packet
timeval to calculate time elapsed.

Conflicts:
	lib/rec.c